fix: corregir Double unitPrice a BigDecimal unitPrice

+++ backend/orders/src/main/java/com/hub/pedidos360/orders/service/OrderService.java
@@ -44,7 +44,7 @@ public class OrderService {
             OrderItem item = new OrderItem();
             item.setProductName(itemRequest.getProductName());
             item.setQuantity(itemRequest.getQuantity());
-            item.setUnitPrice(BigDecimal.valueOf(itemRequest.getUnitPrice()));
+            item.setUnitPrice(itemRequest.getUnitPrice());
             item.setOrder(order);
 
             order.getItems().add(item);
